+++ app/src/main/java/com/adyen/android/assignment/nearbyplaces/domain/model/PlaceListItem.kt
@@ -2,6 +2,6 @@ package com.adyen.android.assignment.nearbyplaces.domain.model
 
 data class PlaceListItem(
     val id: Int?,
-    val distance: Double?,
+    val distance: Int?,
     val name: String?
 )
